cnn-service/requirements.txt
flask==3.0.3
flask-cors==4.0.1
tensorflow==2.16.1
tensorflow-intel==2.16.1
numpy==1.26.4
opencv-python==4.9.0.80
pillow==10.3.0
pytesseract==0.3.10
pdf2image==1.17.0
scikit-learn==1.5.0
matplotlib==3.9.0
requests==2.32.3
werkzeug==3.0.3
